Выделяет дочерний презентер для фильма
