+++ src/presenter/board-presenter.js
@@ -1,15 +1,13 @@
 import { FilmCardsOnPage, FilterType, SortType } from '../consts';
 import { remove, render } from '../framework/render';
-import CommentsModel from '../model/comments-model';
-import SiteFilmCardView from '../view/site-film-card/site-film-card-view';
 import SiteFilmListContainerView from '../view/site-film-list-container/site-film-list-container-view';
 import SiteFilmsListView from '../view/site-film-list/site-films-list-view';
-import SiteFilmPopupView from '../view/site-film-popup/site-film-popup-view';
 import SiteFilmsContainerView from '../view/site-films-container/site-films-container-view';
 import SiteFiltersView from '../view/site-filters/site-filters-view';
 import NoFilmsTemplate from '../view/site-no-films/site-no-films-vies';
 import ShowMoreButtonView from '../view/site-show-more-button/site-show-more-button-view';
 import SiteSortView from '../view/site-sort/site-sort-view';
+import FilmPresenter from './film-presenter';
 
 export default class BoardPresenter {
   #boardContainer = null;
@@ -54,7 +52,6 @@ export default class BoardPresenter {
     render(this.#filmsContainerComponent, this.#boardContainer);
 
     const films = this.films;
-    // const films = [];
     const filmsCount = films.length;
 
     if (filmsCount === 0) {
@@ -78,42 +75,11 @@ export default class BoardPresenter {
   }
 
   #renderFilm (film) {
-    let filmPopup = null;
-
-    const escKeyDownHandler = (evt) => {
-      if (evt.key === 'Escape') {
-        evt.preventDefault();
-        removePopup();
-        document.removeEventListener('keydown', escKeyDownHandler);
-      }
-    };
-
-    function removePopup() {
-      document.body.classList.remove('hide-overflow');
-      remove(filmPopup);
-    }
-
-    const renderPopup = () => {
-      const commentsModel = new CommentsModel(film.id);
-      commentsModel.init().finally(() => {
-        const comments = commentsModel.comments;
-
-        filmPopup = new SiteFilmPopupView(film, comments);
-        filmPopup.setClosePopupHandler(removePopup);
-        render(filmPopup, document.body);
-      });
-    };
-
-    const filmComponent = new SiteFilmCardView({
-      film,
-      onFilmCardClick: () => {
-        renderPopup();
-        document.addEventListener('keydown', escKeyDownHandler);
-        document.body.classList.add('hide-overflow');
-      }
+    const filmPresenter = new FilmPresenter({
+      filmListContainer: this.#filmListContainerComponent.element,
     });
 
-    render(filmComponent, this.#filmListContainerComponent.element);
+    filmPresenter.init(film);
   }
 
   #renderShowMoreButton () {
